requirements.txt
# Django
django-filter
djangorestframework
django-model-utils

# Mongo
djongo

#env
django-environ

#Documentation
drf_yasg

#Deployment
gunicorn
django_heroku